server/app/requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg==3.2.1
psycopg-binary==3.2.1
sqlparse==0.5.1
typing_extensions==4.12.2
